PointNet2/requirements.txt
# PointNet2 项目依赖
torch==2.0.0
torchvision==0.15.0
torchaudio==2.0.0
numpy==1.24.3
scipy==1.10.1
matplotlib==3.7.1
tqdm==4.65.0
h5py==3.8.0
scikit-learn==1.2.2
plyfile==0.7.4
open3d==0.18.0
pyyaml==6.0
pointnet2-ops
pytorch3d